Mega Code Cleanup
Also removed numpy dependencies in arm.py with math, so hopefully that helps with file size on the RoboRIO
